refactor: move CtxLogger and msgCtx into their own module
Extract the msg-ctx logging concern (renderLogCtx, msgCtx, CtxLogger) out of
the catch-all utils.ts into a dedicated src/ctxLogger.ts, and split its tests
into test/ctxLogger.test.ts. Purely internal — these symbols are not part of
the package's public API (never re-exported from index.ts), so no consumer
impact. Importers now pull CtxLogger from './ctxLogger'.
